requirements.txt
flask
line-bot-sdk
python-dotenv
beautifulsoup4
requests